Refactor pino stuff into one single file
