+++ backend/src/utils/config.ts
@@ -2,9 +2,8 @@ import fp from "fastify-plugin";
 import fs from "fs";
 import path from "path";
 import type { FastifyInstance, FastifyServerOptions } from "fastify";
-import type { PinoLoggerOptions } from "fastify/types/logger";
-import { z, ZodError } from "zod";
-import { GameState } from "../game/game.types";
+import { z } from "zod";
+import { devLoggerConfig, prodLoggerConfig } from "./logger";
 
 const configSchema = z.object({
     BACKEND_PORT: z.coerce
@@ -19,52 +18,6 @@ const configSchema = z.object({
     JWT_SECRET: z.string().min(32, "must be at least 32 characters long"),
     DB_PATH: z.string().min(1, "cannot be empty"),
 });
-
-const formatError = (error: unknown) => {
-    if (error instanceof ZodError) {
-        return {
-            name: "ZodError",
-            issues: error.issues.map((i) => `${i.path}: ${i.message}`),
-        };
-    }
-
-    if (error instanceof Error) {
-        return {
-            name: error.name,
-            message: error.message,
-            stack: error.stack ?? "No stack trace available",
-        };
-    }
-
-    /** Fallback for unknown error */
-    return {
-        message: String(error),
-    };
-};
-
-const formatGameState = (state: GameState) => {
-    return {
-        ballPosition: `x: ${state.ballPosition.x}, y: ${state.ballPosition.y}, z: ${state.ballPosition.z}`,
-        paddle1Position: `x: ${state.paddlePosition["player-1"].x}, y: ${state.paddlePosition["player-1"].y}, z: ${state.paddlePosition["player-1"].z}`,
-        paddle2Position: `x: ${state.paddlePosition["player-2"].x}, y: ${state.paddlePosition["player-2"].y}, z: ${state.paddlePosition["player-2"].z}`,
-        score: `player1: ${state.score.player1}, player2: ${state.score.player2}`,
-    };
-};
-
-const devLoggerConfig: PinoLoggerOptions = {
-    level: "debug", // More detailed logs in dev
-    transport: {
-        target: "pino-pretty",
-        options: {
-            colorize: true, // Enables colors for better readability
-            translateTime: "HH:MM:ss Z", // Formats timestamps
-            ignore: "pid,hostname", // Hides unnecessary fields
-        },
-    },
-    serializers: { err: formatError, gameState: formatGameState },
-};
-
-const prodLoggerConfig: PinoLoggerOptions = { level: "info" };
 
 /** Fastify server options, cannot be changed once instance is created */
 export const appOpts: FastifyServerOptions = {
